fix(issue-72): phase 3 priority 3 - shorten UI and hook long functions to under 150 lines
- useAIChat: extract executeSendMessage to useAIChatExecute.ts
- SearchResults: extract SearchResultCard, SearchResultsLoadingSkeleton, SearchResultsEmptyState
- HeaderSearchBar: extract handleSearchKeyDown and HeaderSearchDropdownContent
- AIChatInput: split into aiChatInputHelpers, useAIChatInput (useAIChatInputChips)
- useImageUploadManager: extract runSingleUpload, editor node helpers, filterImageFiles; unify processImageUpload

Ref: #72
