+++ src/pages/SearchResults.tsx
@@ -1,10 +1,11 @@
 import { useMemo, useEffect } from "react";
 import { useSearchParams, useNavigate } from "react-router-dom";
-import { Search, FileText, Link as LinkIcon } from "lucide-react";
 import Header from "@/components/layout/Header";
 import Container from "@/components/layout/Container";
-import { HighlightedSnippet } from "@/components/search/HighlightedSnippet";
-import { MatchTypeBadge } from "@/components/search/MatchTypeBadge";
+import { SearchResultCard } from "@/components/search/SearchResultCard";
+import type { SearchResultCardItem } from "@/components/search/SearchResultCard";
+import { SearchResultsLoadingSkeleton } from "@/components/search/SearchResultsLoadingSkeleton";
+import { SearchResultsEmptyState } from "@/components/search/SearchResultsEmptyState";
 import { useSearchPages, useSearchSharedNotes } from "@/hooks/usePageQueries";
 import { extractPlainText } from "@/lib/contentUtils";
 import {
@@ -16,18 +17,9 @@ import {
   calculateEnhancedScore,
 } from "@/lib/searchUtils";
 import { useGlobalSearchContext } from "@/contexts/GlobalSearchContext";
-import { cn } from "@/lib/utils";
 
-interface SearchResultItem {
-  pageId: string;
-  noteId?: string;
-  title: string;
+interface SearchResultItem extends SearchResultCardItem {
   snippet: string;
-  highlightedSnippet: string;
-  matchType: MatchType;
-  sourceUrl?: string;
-  thumbnailUrl?: string;
-  updatedAt: number;
   score: number;
 }
 
@@ -37,20 +29,15 @@ export default function SearchResults() {
   const { setQuery } = useGlobalSearchContext();
   const searchQuery = (searchParams.get("q") ?? "").trim();
 
-  // ヘッダー検索バーの入力欄と同期
   useEffect(() => {
-    if (searchQuery) {
-      setQuery(searchQuery);
-    }
+    if (searchQuery) setQuery(searchQuery);
   }, [searchQuery, setQuery]);
 
   const { data: personalResults = [], isLoading: isPersonalLoading } = useSearchPages(searchQuery);
   const { data: sharedResponse, isLoading: isSharedLoading } = useSearchSharedNotes(searchQuery);
 
   const isLoading = isPersonalLoading || isSharedLoading;
-
   const keywords = useMemo(() => parseSearchQuery(searchQuery), [searchQuery]);
-
   const sharedResults = useMemo(() => sharedResponse?.results ?? [], [sharedResponse]);
 
   const results = useMemo((): SearchResultItem[] => {
@@ -106,22 +93,14 @@ export default function SearchResults() {
     }
   };
 
-  const formatDate = (ts: number) => {
-    if (!ts) return "";
-    const d = new Date(ts);
-    return d.toLocaleDateString("ja-JP", {
-      year: "numeric",
-      month: "short",
-      day: "numeric",
-    });
-  };
+  const resultKey = (item: SearchResultItem) =>
+    item.noteId ? `shared-${item.noteId}-${item.pageId}` : `personal-${item.pageId}`;
 
   return (
     <div className="min-h-screen bg-background">
       <Header />
       <main className="py-6">
         <Container>
-          {/* 見出し */}
           <div className="mb-6">
             {searchQuery ? (
               <h1 className="text-lg font-medium">
@@ -139,112 +118,33 @@ export default function SearchResults() {
             )}
           </div>
 
-          {/* ローディング */}
-          {isLoading && searchQuery.length >= 3 && (
-            <div className="max-w-3xl space-y-4">
-              {[...Array(3)].map((_, i) => (
-                <div
-                  key={i}
-                  className="flex animate-pulse gap-4 rounded-lg border border-border p-4"
-                >
-                  <div className="h-16 w-24 shrink-0 rounded bg-muted" />
-                  <div className="flex-1 space-y-2">
-                    <div className="h-5 w-1/3 rounded bg-muted" />
-                    <div className="h-4 w-full rounded bg-muted" />
-                    <div className="h-4 w-2/3 rounded bg-muted" />
-                  </div>
-                </div>
-              ))}
-            </div>
-          )}
+          {isLoading && searchQuery.length >= 3 && <SearchResultsLoadingSkeleton />}
 
-          {/* 検索クエリが短すぎる */}
           {searchQuery.length > 0 && searchQuery.length < 3 && (
-            <div className="py-12 text-center">
-              <Search className="mx-auto mb-4 h-12 w-12 text-muted-foreground/40" />
-              <p className="text-muted-foreground">3文字以上入力してください</p>
-            </div>
+            <SearchResultsEmptyState description="3文字以上入力してください" />
           )}
 
-          {/* 結果なし */}
           {!isLoading && searchQuery.length >= 3 && results.length === 0 && (
-            <div className="py-12 text-center">
-              <Search className="mx-auto mb-4 h-12 w-12 text-muted-foreground/40" />
-              <p className="mb-1 text-lg font-medium">検索結果が見つかりません</p>
-              <p className="text-sm text-muted-foreground">別のキーワードで検索してみてください</p>
-            </div>
+            <SearchResultsEmptyState
+              title="検索結果が見つかりません"
+              description="別のキーワードで検索してみてください"
+            />
           )}
 
-          {/* 検索結果リスト */}
           {!isLoading && results.length > 0 && (
             <div className="max-w-3xl space-y-3">
               {results.map((item) => (
-                <button
-                  key={
-                    item.noteId ? `shared-${item.noteId}-${item.pageId}` : `personal-${item.pageId}`
-                  }
-                  type="button"
+                <SearchResultCard
+                  key={resultKey(item)}
+                  item={item}
                   onClick={() => handleResultClick(item)}
-                  className={cn(
-                    "w-full rounded-lg border border-border p-4 text-left",
-                    "transition-colors hover:border-muted-foreground/30 hover:bg-muted/50",
-                    "focus:outline-none focus:ring-2 focus:ring-ring",
-                  )}
-                >
-                  <div className="flex gap-4">
-                    {/* サムネイル */}
-                    {item.thumbnailUrl && (
-                      <div className="shrink-0">
-                        <img
-                          src={item.thumbnailUrl}
-                          alt=""
-                          className="h-20 w-28 rounded bg-muted object-cover"
-                          loading="lazy"
-                        />
-                      </div>
-                    )}
-
-                    <div className="min-w-0 flex-1">
-                      {/* タイトル行 */}
-                      <div className="mb-1.5 flex items-center gap-2">
-                        {item.sourceUrl ? (
-                          <LinkIcon className="h-4 w-4 shrink-0 text-muted-foreground" />
-                        ) : (
-                          <FileText className="h-4 w-4 shrink-0 text-muted-foreground" />
-                        )}
-                        <h3 className="flex-1 truncate text-base font-medium">{item.title}</h3>
-                        <MatchTypeBadge type={item.matchType} />
-                        {item.noteId && (
-                          <span className="shrink-0 rounded bg-blue-100 px-1.5 py-0.5 text-[10px] text-blue-700 dark:bg-blue-900 dark:text-blue-300">
-                            共有
-                          </span>
-                        )}
-                      </div>
-
-                      {/* ハイライト付きスニペット */}
-                      <div className="mb-1.5">
-                        <HighlightedSnippet text={item.highlightedSnippet} />
-                      </div>
-
-                      {/* メタ情報 */}
-                      <div className="flex items-center gap-3 text-xs text-muted-foreground">
-                        <span>{formatDate(item.updatedAt)}</span>
-                      </div>
-                    </div>
-                  </div>
-                </button>
+                />
               ))}
             </div>
           )}
 
-          {/* 初期状態（クエリ未入力） */}
           {!searchQuery && (
-            <div className="py-12 text-center">
-              <Search className="mx-auto mb-4 h-12 w-12 text-muted-foreground/40" />
-              <p className="text-muted-foreground">
-                ヘッダーの検索バーからキーワードを入力してください
-              </p>
-            </div>
+            <SearchResultsEmptyState description="ヘッダーの検索バーからキーワードを入力してください" />
           )}
         </Container>
       </main>
